用户登录后的ws连接处理，加载离线时的消息列表、离线时的申请列表、联系人会话等，并增加用户断开连接时的离线时间，以及调整一些字段类型

+++ chat-service/src/main/java/com/salmon/chatService/service/impl/GroupServiceImpl.java
@@ -155,8 +155,8 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, Group> implements
      */
     @Override
     public GroupChatVO getGroupInfo4Chat(IdRequest request) {
-        Long groupId = request.getId();
-        Group group = this.getGroupDetail(groupId.intValue());
+        Integer groupId = request.getId().intValue();
+        Group group = this.getGroupDetail(groupId);
         GroupVO groupVO = GroupVO.objToVo(group);
         List<UserContactVO> userContactVOS = userContactService.selectContactUserInfo(groupId, UserContactTypeEnum.GROUP.getType(), UserContactStatusEnum.FRIEND.getValue());
         if (CollectionUtils.isEmpty(userContactVOS)) {
